+++ packages/patrol_devtools_extension/lib/native_inspector/nodes/node.dart
@@ -4,6 +4,7 @@ import 'package:patrol_devtools_extension/api/native_views.dart';
 part 'android_node.dart';
 part 'ios_node.dart';
 part 'native_view_node.dart';
+part 'web_node.dart';
 
 sealed class Node {
   List<Node> get children;